==> 1_assign_tweets_topic_distribution.ipynb <==
import pandas as pd
import numpy as np
import gensim

from helper.data_loading import *
from helper.preprocessing import *
from helper.topic_model_helper import *

pd.set_option('display.max_colwidth', -1)

#load model,dict,create corpus
model = gensim.models.ldamodel.LdaModel.load("output/topic_model/model_lda")
dictionary = gensim.corpora.Dictionary.load("output/topic_model/dict")

### Prepare  data for assigning topic distributions

### GermEval
germEval = get_all_germEval_data()
germEval = assign_topic_distribution(germEval, model, dictionary)
germEval = assign_topic_and_probability(germEval)

germEval.to_pickle("data/GERMEVAL_with_topic_distribution.pkl")

### Left

left = pd.read_pickle("data/bipolarity_crawler-master/final_tweets/dataset_LEFT.pkl")
left = assign_topic_distribution(left, model, dictionary)
left  = assign_topic_and_probability(left)

left.to_pickle("data/LEFT_with_topic_distribution.pkl")

### Right

right = pd.read_pickle("data/bipolarity_crawler-master/final_tweets/dataset_RIGHT.pkl")
right = assign_topic_distribution(right, model, dictionary)
right = assign_topic_and_probability(right)

right.to_pickle("data/RIGHT_with_topic_distribution.pkl")

### Neutral

neutral = pd.read_pickle("data/bipolarity_crawler-master/final_tweets/dataset_NEUTRAL.pkl")
neutral = assign_topic_distribution(neutral, model, dictionary)
neutral = assign_topic_and_probability(neutral)

neutral.to_pickle("data/NEUTRAL_with_topic_distribution.pkl")

pd.set_option('display.max_colwidth', -1)

neutral = pd.read_pickle("data/NEUTRAL_with_topic_distribution.pkl")

neutral.head()
