==> 4_plot.ipynb <==
import numpy as np
import pandas as pd
import pickle
import statistics
from scipy import stats


from matplotlib import pyplot as plt
import numpy as np
from matplotlib.lines import Line2D   

import matplotlib.colors as mcolors
cols = [color for name, color in mcolors.TABLEAU_COLORS.items()] 

from helper.helper_plots import *

pd.set_option('display.max_colwidth', -1)

left_result,left_ids = get_result_files("output/classifier/left/")
right_result,right_ids = get_result_files("output/classifier/right/")
neutral_result,neutral_ids = get_result_files("output/classifier/neutral/")

left_ids = get_id_dict(left_ids)
right_ids = get_id_dict(right_ids)
neutral_ids = get_id_dict(neutral_ids)

left_result = get_result_dict(left_result)
right_result = get_result_dict(right_result)
neutral_result = get_result_dict(neutral_result)

### Boxplot

fig1, ax1 = plt.subplots(figsize=(15, 8))
ax1.set_title('Left, Neutral and Right F1-scores - 100% data replacement')
ax1.set_xlabel('Datasets')
ax1.set_ylabel('F1-score - Overall')

ax1.boxplot([right_result["overall"][3], left_result["overall"][3], neutral_result["overall"][3]])
ax1.yaxis.grid(color='gray', linestyle='dashed')

plt.xticks([1, 2, 3], ['Right', 'Left',"Neutral"])

plt.show()

fig1.savefig("output/figures/boxplot.pdf")

len(neutral_result["overall"][3])

### Graph - Overall

x = [0.0,0.33,0.66,1]

left_overall_mean, left_overall_stdev = get_statistics(left_result["overall"])
right_overall_mean , right_overall_stdev = get_statistics(right_result["overall"])
neutral_overall_mean , neutral_overall_stdev = get_statistics(neutral_result["overall"])

#plot npmi score
fig, ax = plt.subplots(figsize=(15, 8))

ax.plot(x, left_overall_mean, lw=2, label='Left data', color="#E46C0A")
ax.plot(x, right_overall_mean, lw=2, label='Right data', color="#31859C")
ax.plot(x, neutral_overall_mean, lw=2, label='Neutral data', color="#604A7B")

ax.fill_between(x, left_overall_mean+left_overall_stdev, left_overall_mean-left_overall_stdev, facecolor="#E46C0A", alpha=0.5)
ax.fill_between(x, right_overall_mean+right_overall_stdev, right_overall_mean-right_overall_stdev, facecolor="#31859C", alpha=0.5)
ax.fill_between(x, neutral_overall_mean+neutral_overall_stdev, neutral_overall_mean-neutral_overall_stdev, facecolor="#604A7B", alpha=0.5)

ax.set_title(r'Left, Neutral and Right F1-scores - [0%,33%,66%,100%] data replacement')
ax.legend(loc='upper left')
ax.set_xlabel('Percentage of data replacement')
ax.set_ylabel('F1-score - Overall')
ax.yaxis.grid(color='gray', linestyle='dashed')

plt.show()               

fig.savefig("output/figures/overall_rolling.pdf")

94/7

### Graph - Offense

x = [0.0,0.33,0.66,1]

left_overall_mean, left_overall_stdev = get_statistics(left_result["offense"])
right_overall_mean , right_overall_stdev = get_statistics(right_result["offense"])
neutral_overall_mean , neutral_overall_stdev = get_statistics(neutral_result["offense"])

#plot npmi score
fig, ax = plt.subplots(figsize=(15, 8))

ax.plot(x, left_overall_mean, lw=2, label='Left data', color="#E46C0A")
ax.plot(x, right_overall_mean, lw=2, label='Right data', color="#31859C")
ax.plot(x, neutral_overall_mean, lw=2, label='Neutral data', color="#604A7B")

ax.fill_between(x, left_overall_mean+left_overall_stdev, left_overall_mean-left_overall_stdev, facecolor="#E46C0A", alpha=0.5)
ax.fill_between(x, right_overall_mean+right_overall_stdev, right_overall_mean-right_overall_stdev, facecolor="#31859C", alpha=0.5)
ax.fill_between(x, neutral_overall_mean+neutral_overall_stdev, neutral_overall_mean-neutral_overall_stdev, facecolor="#604A7B", alpha=0.5)

ax.set_title(r'Left, Neutral and Right F1-scores - [0%,33%,66%,100%] data replacement')
ax.legend(loc='upper left')
ax.set_xlabel('Percentage of data replacement')
ax.set_ylabel('F1-score - Offensive')
ax.yaxis.grid(color='gray', linestyle='dashed')

plt.show()               

fig.savefig("output/figures/offense_rolling.pdf")

neutral_overall_mean

statistics.mean(right_overall_mean)

### Graph - None

x = [0.0,0.33,0.66,1]

left_overall_mean, left_overall_stdev = get_statistics(left_result["none"])
right_overall_mean , right_overall_stdev = get_statistics(right_result["none"])
neutral_overall_mean , neutral_overall_stdev = get_statistics(neutral_result["none"])

#plot npmi score
fig, ax = plt.subplots(figsize=(15, 8))

ax.plot(x, left_overall_mean, lw=2, label='Left data', color="#E46C0A")
ax.plot(x, right_overall_mean, lw=2, label='Right data', color="#31859C")
ax.plot(x, neutral_overall_mean, lw=2, label='Neutral data', color="#604A7B")

ax.fill_between(x, left_overall_mean+left_overall_stdev, left_overall_mean-left_overall_stdev, facecolor="#E46C0A", alpha=0.5)
ax.fill_between(x, right_overall_mean+right_overall_stdev, right_overall_mean-right_overall_stdev, facecolor="#31859C", alpha=0.5)
ax.fill_between(x, neutral_overall_mean+neutral_overall_stdev, neutral_overall_mean-neutral_overall_stdev, facecolor="#604A7B", alpha=0.5)

ax.set_title(r'Left, Neutral and Right F1-scores - [0%,33%,66%,100%] data replacement')
ax.legend(loc='upper left')
ax.set_xlabel('Percentage of data replacement')
ax.set_ylabel('F1-score - Non-Offensive')
ax.yaxis.grid(color='gray', linestyle='dashed')

plt.show()               

fig.savefig("output/figures/none_rolling.pdf")

neutral_overall_mean

### Statistical tests

def perform_test(data1,data2):
    shapiro1 = list(map(lambda x: round(x,3), stats.shapiro(data1)))
    shapiro2 = list(map(lambda x: round(x,3), stats.shapiro(data2)))
    levene1 = list(map(lambda x: round(x,3), stats.bartlett(data1, data2)))
    ks_2samp = list(map(lambda x: round(x,3), stats.ks_2samp(data1, data2)))
    print(ks_2samp)
    print("Data Normaly Distributed: data1: {}, data2: {}".format(shapiro1[1],shapiro2[1]))
    print(" Equal Variance: \n  P-Value: {}".format(levene1[1]))
    print(" T-Test:\n  T-Value: {}\n  P-Value: {}\n".format(ks_2samp[0],ks_2samp[1]))

print("Right and Left")
perform_test(right_result["overall"][3],left_result["overall"][3])

print("Right and Neutral")
perform_test(right_result["overall"][3],neutral_result["overall"][3])

print("Left and Neutral")
perform_test(left_result["overall"][3],neutral_result["overall"][3])
